Add dynamic div sizing via JS; remove static ".one" div sizing in CSS
The program now dynamically sizes the ".one" divs to fit within the 800x800 box.
Currently "Math.floor()" is used to keep the div sizes towards whole numbers but
this may change later, as it means some "box resolutions" result in a few pixels
of space between the edge of the divs and the edge of the box.

TODO - User input via UI, push